safari-600.7-branch/LayoutTests/fast/loader/form-state-restore-with-locked-back-forward-list.html

<!DOCTYPE html>
<html>
<body onLoad="document.forms[0].submit();">
<form method=POST action="resources/form-state-restore-with-locked-back-forward-list-2.html">
<input name="TestName1" type="hidden" value="Bad1">
<input name="TestName2" type="hidden" value="Bad2">
</form>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}
</script>
</body>
</html>
